Refactor flash and update to share more code.
